oidc: make oidc_provider dependency optional
