+++ prologin/oidc_policy/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class OidcPolicyConfig(AppConfig):
+    name = 'oidc_policy'
