+++ prologin/users/admin.py
@@ -2,7 +2,6 @@ from django.contrib import admin
 from django.contrib.auth import get_user_model
 from django.contrib.auth.admin import UserAdmin
 from django.utils.translation import ugettext_lazy as _
-from .models import OpenIDClientPolicy
 
 
 class ProloginUserAdmin(UserAdmin):
@@ -14,21 +13,5 @@ class ProloginUserAdmin(UserAdmin):
                                     'preferred_language',)}),
     )
 
-class OpenIDClientPolicyAdmin(admin.ModelAdmin):
-    search_fields = ('openid_client__name',)
-    list_display = (
-        'openid_client',
-        'allow_staff',
-        'allow_assigned_semifinal',
-        'allow_assigned_semifinal_event',
-        'allow_assigned_final',
-    )
-    list_filter = (
-        'allow_staff',
-        'allow_assigned_final',
-        'allow_assigned_semifinal',
-        'allow_groups',
-    )
 
 admin.site.register(get_user_model(), ProloginUserAdmin)
-admin.site.register(OpenIDClientPolicy, OpenIDClientPolicyAdmin)
